+++ src/app/app-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { HomeComponent } from './home/home.component';
-import { AboutComponent } from './about/about.component';
-import { SkillsComponent } from './skills/skills.component';
-import { ProjectsPageComponent } from './projects/projects-page/projects-page.component';
-import { Project1Component } from './projects/project1/project1.component';
-import { Project2Component } from './projects/project2/project2.component';
-import { Project3Component } from './projects/project3/project3.component';
+import { HomeComponent } from './components/home/home.component';
+import { AboutComponent } from './components/about/about.component';
+import { SkillsComponent } from './components/skills/skills.component';
+import { ProjectsPageComponent } from './components/projects/projects-page/projects-page.component';
+import { Project1Component } from './components/projects/project1/project1.component';
+import { Project2Component } from './components/projects/project2/project2.component';
+import { Project3Component } from './components/projects/project3/project3.component';
 
 const routes: Routes = [
   {
