+++ src/app/components/projects/project1/project1.component.html
@@ -19,8 +19,7 @@
     <p>
       <span>April 2019 - June 2019</span>
       <br />
-      {{ mindstorm }}
-      <br />
+      <span>{{ mindstorm }}</span>
       <br />
       <a
         target="_blank"
