+++ src/app/components/projects/project1/project1.component.ts
