+++ src/app/components/projects/project2/project2.component.ts
