+++ src/app/shared/skills.model.ts
